Added utils dir and split up utils
