+++ src/print_structs.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   print_structs.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: hrother <hrother@student.42vienna.com>     +#+  +:+       +#+        */
+/*   By: aprevrha <aprevrha@student.42vienna.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/09 22:12:46 by hrother           #+#    #+#             */
-/*   Updated: 2024/04/30 18:03:22 by hrother          ###   ########.fr       */
+/*   Updated: 2024/05/06 21:01:55 by aprevrha         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -46,19 +46,18 @@ void	print_token(void *token)
 void	print_token_new(void *p_tkn)
 {
 	t_token	*token;
+	char	*names[9];
 
+	names[0] = "NOTDEF		";
+	names[1] = "CMD			";
+	names[2] = "ARG			";
+	names[3] = "R_IN		";
+	names[4] = "R_OUT		";
+	names[5] = "R_APPEND	";
+	names[6] = "R_HEREDOC	";
+	names[7] = "PIPE		";
+	names[8] = "R_QUOTEDOC	";
 	token = ((t_token *)p_tkn);
-	char *names[] = {
-		"NOTDEF		",
-		"CMD		",
-		"ARG		",
-		"R_IN		",
-		"R_OUT		",
-		"R_APPEND	",
-		"R_HEREDOC	",
-		"PIPE		",
-		"R_QUOTEDOC	",
-	};
 	if (token->type < 0 || token->type > 8)
 		log_msg(WARNING, "type = %i: Type out of range!\n", token->type);
 	log_msg(DEBUG, "%i:%s [%s]", token->type, names[token->type], token->value);
