+++ src/utils/utils.c
@@ -1,0 +1,43 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   utils.c                                            :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: aprevrha <aprevrha@student.42vienna.com    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/02/28 16:41:35 by hrother           #+#    #+#             */
+/*   Updated: 2024/05/06 20:51:10 by aprevrha         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../../includes/minishell.h"
+
+int	ft_strcmp(const char *s1, const char *s2)
+{
+	return (ft_strncmp(s1, s2, ft_strlen(s2) + 1));
+}
+
+t_cmd	*new_cmd(char *bin, char **args)
+{
+	t_cmd	*cmd;
+
+	cmd = (t_cmd *)malloc(sizeof(t_cmd));
+	if (cmd == NULL)
+		return (log_msg(ERROR, "malloc failed"), NULL);
+	cmd->bin = bin;
+	cmd->args = args;
+	cmd->redirects = NULL;
+	cmd->fd_in = STDIN_FILENO;
+	cmd->fd_out = STDOUT_FILENO;
+	cmd->pid = -1;
+	cmd->status = 0;
+	return (cmd);
+}
+
+/**
+ * @brief This function does nothing: Useful for ft_lstclear
+ */
+void	pass(void *content)
+{
+	(void)content;
+}
