+++ src/utils/utils_free.c
@@ -1,38 +1,16 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   utils.c                                            :+:      :+:    :+:   */
+/*   utils_free.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: hrother <hrother@student.42vienna.com>     +#+  +:+       +#+        */
+/*   By: aprevrha <aprevrha@student.42vienna.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/28 16:41:35 by hrother           #+#    #+#             */
-/*   Updated: 2024/04/30 18:03:22 by hrother          ###   ########.fr       */
+/*   Updated: 2024/05/06 20:51:23 by aprevrha         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../includes/minishell.h"
-
-int	ft_strcmp(const char *s1, const char *s2)
-{
-	return(ft_strncmp(s1, s2, ft_strlen(s2) + 1));
-}
-
-t_cmd	*new_cmd(char *bin, char **args)
-{
-	t_cmd	*cmd;
-
-	cmd = (t_cmd *)malloc(sizeof(t_cmd));
-	if (cmd == NULL)
-		return (log_msg(ERROR, "malloc failed"), NULL);
-	cmd->bin = bin;
-	cmd->args = args;
-	cmd->redirects = NULL;
-	cmd->fd_in = STDIN_FILENO;
-	cmd->fd_out = STDOUT_FILENO;
-	cmd->pid = -1;
-	cmd->status = 0;
-	return (cmd);
-}
+#include "../../includes/minishell.h"
 
 void	free_token(void *content)
 {
@@ -76,10 +54,3 @@ void	free_nullterm_str_arr(char **strs)
 	free(strs);
 }
 
-/**
- * @brief This function does nothing: Useful for ft_lstclear
- */
-void	pass(void *content)
-{
-	(void)content;
-}
